Add files via upload
This is the first revision of the timing development program. There have been some tweaks to the variables and the code from the original graphing program. This is partly due to the addition of scales on the left & right hand side, and the addition of margins in which to display data.