Add files via upload
The program has now been divided into several tabs for clarity. The COM port is now automatically assigned to the highest open port. The timings are monitored and displayed for a single hard-defined pressure window; though the code is written using an array to allow for 3 pressure windows. There are still some graphical bugs which can be ironed out, these don't significantly affect the functionality of the program.