Add files via upload
the OBJECTIVES achieved are:
1) Allow 3rd timing window to be below the plot & so out of use
2) Create a calibration system for calibrating the input from a real transducer

A calibration procedure has been created which prompts the user for any 2 pressures to be calibrated, preferably maximally distant. It then suggests a scale which is slightly wider than the range of the calibration, but this can be adjusted with free type by the user. Currently all calibrations are in cmH2O but I plan to add other units, and also plan to be able to save or load up to 4 profiles.