Add files via upload
I've added a screen to allow the user to choose the pressure units at the start of calibration. At the end the user is asked to select a profile to save the calibration to, or to continue without saving. Presently nothing is saved but the button selection is noted so actions can be taken in the future.
I've added gridlines to the plot to make it easier to see the pressure at any time during the trace.
I've increased the stroke weight (thickness) of the trace to make it easier to see.
I've added additional plot frames within a single draw frame e.g., progress the plot in 4x 0.25 horizontal pixel increments within a single draw frame. The advantage of this is that it gives a better trace when the output is swinging quickly.