Add files via upload
I’ve now added the actual save & load data functions. This included the need to change several key variables relating to the raw digital data, scales, and pressure units to be arrays, holding 4 data sets. A 5th dataset is used to temporarily store the present calibration data, if it is to be saved it is copied to the selected data set.