Add files via upload
When attempting to confirm the timing accuracy of the system I found that when confronted with a near pure square-wave, some of the vertical lines weren't being drawn. I revisited the dub-frame creation portion of the code and made substantial changes. The program now has both start and end X and Y coordinates, and the difference between the start and end is divided by the number of sub-frames to yield an increment factor. Then multiple lines are drawn at these much smaller increments. Furthermore, the decision of what colour to draw the lines is based on the current Y coordinate of each sub-line rather than the the status of the pressure window which updates at the same rate as the master frame rate.