Add files via upload
Configured the Arduino to read & transmit both data channels.
Amended the program to read both data channels and split the data into 2 readings.
Amended the existing saved configs to include the low / high pressure channel i.e., 0 or 1, then amended the program to load and save the extra pressure channel data. Also added a step in the calibration routine to allow the user to select the low / high range pressure channel.
Made the time-base button active and created 6 different pre-sets: 5s, 10s, 15s, 20s, 30s, 60s/
Made a function to test exclusivity of button selections as there are now 25 buttons. This works by passing in the button number of the button of interest and checking if any others are flagged as being held (when one pressed the button and then moves the mouse off it).
Increased the plot width to accommodate an extra button above it to load and select a config.