Add files via upload
Running the program on the laptop presented a couple of issues with the display.
Firstly because the 1366x768 frame doesn't include the compulsory program title bar and close buttons, therefore the bottom of the program was off the bottom of the screen. This has been fixed by altering the program window size and slightly adjusting the alignment of program components such as the buttons.
Secondly, though the timing was being measured correctly, the plot trace didn't align with the scale, 10s would take up roughly 6s for example. I realised that this was a refresh rate problem. On the PC I run the program at 100fps, above the monitor refresh rate but effectively let Windows & Java deal with that issue. On the laptop however, it seems that the refresh of the monitor affects the refresh of the program and so with a 60 fps cap, the program was advancing only 60% of the horizontal (time axis) distance across the screen for any given period. This has been fixed by setting the frame rate of the program to 60Hz, having no obvious detriment to the program as seen on the PC, but fixing it on the laptop.
Thirdly, an annoying line which flashes on screen at the start of the program was also appearing when I changed time base whilst a square wave signal was applied. After selecting the new time base, effectively a new trace is drawn on a blank plot, but this almost always started with a vertical line out of position, just like at the start of the program. Both problems were caused by a mismatch of data between the current and old Y positions. When the program starts there is no old Y position so it draws a line from zero to the current position; when a new time base has been selected and a new trace drawn, the old Y position is from seconds earlier, and doesn't marry up with the current Y position, so if you were on a positive phase and are now on a negative phase (or vice versa) then a vertical line is drawn. I have utilised a logic trigger called newLine which addresses the issue in both cases.