Add files via upload
Prior to verifying pressure accuracy, and from observations using the program, I have dampened the refresh rate of the live pressure display. It is customisable in a variable within the source code (but not from within the program) and at default is set to 3 per second. I also scaled the live pressure reading to #.xx if the highest absolute scale magnitude is <10, or to #.x between 10 & 100, and to just # (no decimal places) above 100.
Finally, I took the opportunity to fix a minor bug in the timing readout; if the pressure was in-range but waiting for a new cycle before starting the timing, when the pressure left the current window, the total time the program had waited with the pressure in-range would be displayed in that timing window until overwritten. It did not contribute to any real totals or averages.