Add files via upload
There were issues with loading & saving data sets and when calibrating. I've set the calibration buttons as the default units regardless of the current 4 profiles which could all have the same units as each other. I've fixed an issue where if you saved a high-range scale adjustment or full calibration over the top of a low-range dataset then the long range would be selected. Fairly sure it is now working OK.